Various improvements.
Add the option for the user to specify which landmarks to use for the face rectangle.
Make it check that the face rectangle contains any of the hand landmarks
instead of checking if the rectangles overlap.
Add the option for the user to specify which hand landmarks should be
taken into account.